+++ CHANGELOG.md
@@ -1,6 +1,34 @@
 # Changelog
 
 
+## [14.9.1](https://github.com/blackbaud/skyux/compare/14.9.0...14.9.1) (2026-07-31)
+
+
+### Bug Fixes
+
+* **components/charts:** fix horizontal bar spacing; add visual tests ([#4552](https://github.com/blackbaud/skyux/issues/4552)) ([f33971f](https://github.com/blackbaud/skyux/commit/f33971feeca2e759313ef9a39ec286e509a34217))
+
+## [14.9.0](https://github.com/blackbaud/skyux/compare/14.8.0...14.9.0) (2026-07-29)
+
+
+### Features
+
+* **components/charts:** add code examples ([#4547](https://github.com/blackbaud/skyux/issues/4547)) ([ea6c379](https://github.com/blackbaud/skyux/commit/ea6c37937807f4fd99ccf90a5ef6daf378b1baf1)), closes [AB#4027719](https://dev.azure.com/blackbaud/Products/_workitems/edit/4027719)
+
+## [14.8.0](https://github.com/blackbaud/skyux/compare/14.7.0...14.8.0) (2026-07-28)
+
+
+### Features
+
+* **components/charts:** add bar chart component (developer preview) ([#4527](https://github.com/blackbaud/skyux/issues/4527)) ([c4ddd62](https://github.com/blackbaud/skyux/commit/c4ddd62c829e4a103b02963d1fc8774de00a5063)), closes [AB#4027719](https://dev.azure.com/blackbaud/Products/_workitems/edit/4027719)
+* **components/charts:** add testing harnesses ([#4542](https://github.com/blackbaud/skyux/issues/4542)) ([61c3cca](https://github.com/blackbaud/skyux/commit/61c3cca649410ec45f0964cc75aec1c3f3949f21)), closes [AB#4027719](https://dev.azure.com/blackbaud/Products/_workitems/edit/4027719)
+* **sdk/skyux-stylelint:** update stylelint dependency range ([#4548](https://github.com/blackbaud/skyux/issues/4548)) ([7a1ea2d](https://github.com/blackbaud/skyux/commit/7a1ea2d3fad6ece0c3bb37efcdf11458eb10b13a)), closes [AB#4068309](https://dev.azure.com/blackbaud/Products/_workitems/edit/4068309)
+
+
+### Bug Fixes
+
+* **components/layout:** tokenize page summary background color ([#4536](https://github.com/blackbaud/skyux/issues/4536)) ([ae6b0a6](https://github.com/blackbaud/skyux/commit/ae6b0a6c8fa3f8cd4a82ee185dc159fe54d4a6fe))
+
 ## [15.0.0-alpha.2](https://github.com/blackbaud/skyux/compare/15.0.0-alpha.1...15.0.0-alpha.2) (2026-07-22)